requirements.txt
discum
discord
colorama
mega.py
pyinstaller